refactor: move types and interfaces to a different folder
